+++ CHANGELOG.md
@@ -18,6 +18,7 @@
     - Added asString method to the Data Type helper class
     - Added check for NaN labels in Labeled dataset
     - Changed namespace of Data Type helper
+    - Numeric String Converter now handles NaN strings
 
 - 0.0.15-beta
     - Added Gaussian MLE anomaly detector